fix(ndarray,iter,ufunc): add overlap detection, readonly tracking, and reduction edge cases
NdLayout (fnp-ndarray):
- Add writeable and has_internal_overlap fields to NdLayout for
  accurate view semantics matching NumPy behavior
- Implement detect_internal_overlap() to identify aliased indices
  in strided views using offset set analysis
- Enforce readonly propagation: broadcast views and sliding windows
  are automatically marked read-only

Iterator (fnp-iter):
- Fix overlap_copy_policy to accept zero-length transfers as NoCopy
  instead of erroring — zero-length copies are valid no-ops
- Update test expectations: rename overlap_rejected fixtures to
  overlap_is_no_copy reflecting corrected semantics

UFuncArray (fnp-ufunc):
- Fix any/all reductions over empty axes to return correct identity
  values (false for any, true for all)

Random (fnp-random):
- Fix type casting in binomial sampling (usize→i64 conversions)

Conformance:
- Expand oracle test coverage (datetime: 22→29, dtype: 25→30)
- Update adversarial case fixtures for corrected overlap behavior

+++ crates/fnp-random/src/lib.rs
@@ -2839,9 +2839,9 @@ impl Generator {
                         break;
                     }
                     let p_cond = (p / p_remaining).clamp(0.0, 1.0);
-                    let draws = self.sample_binomial_single(remaining as u64, p_cond, &mut cache);
+                    let draws = self.sample_binomial_single(remaining, p_cond, &mut cache);
                     result[i] = draws as u64;
-                    remaining -= draws as i64;
+                    remaining -= draws;
                     p_remaining -= p;
                     if p_remaining <= 0.0 {
                         p_remaining = 1e-15;
